Fix or suppress 1-instance findbug warnings.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1616312 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/layoutmgr/inline/AlignmentContext.java
@@ -318,13 +318,13 @@ public class AlignmentContext implements Constants {
             case EN_BASELINE: //Nothing to do
                 break;
             case EN_SUB:
-                baselineShiftValue = Math.round(-(xHeight / 2)
-                                + parentAlignmentContext.getActualBaselineOffset(EN_ALPHABETIC)
+                baselineShiftValue = Math.round(-((float)xHeight / 2)
+                                + (float)parentAlignmentContext.getActualBaselineOffset(EN_ALPHABETIC)
                                 );
                 break;
             case EN_SUPER:
-                baselineShiftValue = Math.round(parentAlignmentContext.getXHeight()
-                                + parentAlignmentContext.getActualBaselineOffset(EN_ALPHABETIC)
+                baselineShiftValue = Math.round((float)parentAlignmentContext.getXHeight()
+                                + (float)parentAlignmentContext.getActualBaselineOffset(EN_ALPHABETIC)
                                 );
                 break;
             case 0: // A <length> or <percentage> value
